requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
sentence-transformers==3.3.1
fastapi-mcp
psycopg2-binary
python-dotenv